+++ package.json
@@ -13,7 +13,6 @@
   "keywords": [
     "weather",
     "tide",
-    "react",
     "parcel"
   ],
   "author": "RomanKP",
@@ -22,18 +21,12 @@
     "url": "https://github.com/romankp/weather-station/issues"
   },
   "homepage": "https://github.com/romankp/weather-station#readme",
-  "browserslist": "last 1 Chrome version, last 1 Firefox version",
+  "browserslist": "last 2 Chrome version, last 2 Firefox version",
   "devDependencies": {
     "@babel/core": "7.16",
     "@babel/preset-env": "7.16",
-    "@babel/preset-react": "7.16",
-    "@parcel/transformer-less": "2.0.1",
+    "@parcel/transformer-less": "2.2",
     "less": "4.1",
-    "parcel": "2.0.1"
-  },
-  "dependencies": {
-    "prop-types": "15.8",
-    "react": "17.0",
-    "react-dom": "17.0"
+    "parcel": "2.2"
   }
 }
