--- src/components/Root.less
@@ -1,9 +1,0 @@
-.wrapper {
-  position: relative;
-  max-width: 700px;
-  opacity: 0;
-  transition: opacity 0.5s ease;
-  &.show {
-    opacity: 1;
-  }
-}
