--- src/index.less
@@ -1,13 +1,0 @@
-@import (css)
-  url('https://fonts.googleapis.com/css2?family=Quicksand:wght@300;400&family=Source+Sans+Pro:ital,wght@0,400;0,700;1,400&display=swap');
-
-@import './style/constants.less';
-@import './style/normalizer.less';
-
-// Base
-#root {
-  min-height: 100vh;
-}
-
-// Components
-@import './components/Root.less';
