+++ src/style.less
@@ -1,0 +1,31 @@
+@import (css)
+  url('https://fonts.googleapis.com/css2?family=Quicksand:wght@300;400&family=Source+Sans+Pro:ital,wght@0,400;0,700;1,400&display=swap');
+
+@import './style/constants.less';
+@import './style/normalizer.less';
+
+// Base
+// #root {
+//   color: blue;
+//   min-height: 100vh;
+//   position: relative;
+//   max-width: 700px;
+//   // opacity: 0;
+//   // transition: opacity 0.5s ease;
+//   &.show {
+//     opacity: 1;
+//   }
+// }
+
+// Base
+app-root {
+  // color: blue;
+  min-height: 100vh;
+  position: relative;
+  max-width: 700px;
+  // opacity: 0;
+  // transition: opacity 0.5s ease;
+  &.show {
+    opacity: 1;
+  }
+}
