+++ src/style/normalizer.less
@@ -1,5 +1,9 @@
-* {
+// :root {
+//   --norm-spacing: 0;
+//   --norm-border: none;
+// }
+
+body {
   margin: 0;
   padding: 0;
-  border: none;
 }
